hide k12 behind the feature gate
