+++ .gitignore
@@ -2,3 +2,4 @@ target
 Cargo.lock
 *.swp
 .idea
+.DS_Store
