+++ Cargo.toml
@@ -4,7 +4,7 @@ description = "An implementation of the FIPS-202-defined SHA-3 and SHAKE functio
 homepage = "https://github.com/debris/tiny-keccak"
 license = "CC0-1.0"
 name = "tiny-keccak"
-version = "1.4.4"
+version = "1.5.0"
 authors = ["debris <marek.kotewicz@gmail.com>"]
 keywords = ["sha3", "sha-3", "keccak", "crypto", "kangarootwelve"]
 categories = ["cryptography", "no-std"]
@@ -25,3 +25,24 @@ members = ["comparison"]
 
 [badges]
 travis-ci = { repository = "debris/tiny-keccak", branch = "master"}
+
+[features]
+default = ["keccak"]
+keccak = []
+k12 = []
+
+[[test]]
+name = "keccak"
+required-features = ["keccak"]
+
+[[test]]
+name = "kangaroo"
+required-features = ["k12"]
+
+[[bench]]
+name = "keccak"
+required-features = ["keccak"]
+
+[[bench]]
+name = "kangaroo"
+required-features = ["k12"]
