+++ benches/keccak.rs
@@ -1,14 +1,8 @@
-//! should be started with:
-//! ```bash
-//! multirust run nightly cargo bench
-//! ```
-
 #![feature(test)]
 
 extern crate test;
 
 use tiny_keccak::*;
-
 use test::Bencher;
 
 #[bench]
@@ -44,17 +38,5 @@ fn bench_keccak256(b: &mut Bencher) {
         let data = [0u8; BYTES];
         let mut result = [0u8; BYTES];
         Keccak::keccak256(&data, &mut result);
-    });
-}
-
-#[bench]
-fn bench_k12(b: &mut Bencher) {
-    const BYTES: usize = 32;
-    b.bytes = BYTES as u64;
-
-    b.iter(|| {
-        let data = [0u8; BYTES];
-        let mut result = [0u8; BYTES];
-        k12(&[], &data, &mut result);
     });
 }
